fix wasm examples import after crate rename
