+++ sdk/bindings/wasm/examples/04-migrate_wallet_from_mnemonic.ts
@@ -1,4 +1,4 @@
-import * as wasm from "../pkg/cryptpay_sdk_wasm";
+import * as wasm from "../pkg/cawaena_sdk_wasm";
 import { initSdk } from './utils';
 
 async function main() {
