+++ sdk/bindings/wasm/examples/17-create_customer.ts
@@ -1,4 +1,4 @@
-import * as wasm from "../pkg/cryptpay_sdk_wasm";
+import * as wasm from "../pkg/cawaena_sdk_wasm";
 import { initSdk } from './utils';
 
 async function main() {
